+++ src/main/java/de/codescape/jira/plugins/scrumpoker/data/DefaultStoryPointSupport.java
@@ -7,15 +7,10 @@ import com.atlassian.jira.issue.MutableIssue;
 import com.atlassian.jira.issue.fields.CustomField;
 import com.atlassian.jira.security.JiraAuthenticationContext;
 import com.atlassian.jira.user.ApplicationUser;
-import com.atlassian.jira.util.ErrorCollection;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.text.NumberFormat;
 
 public class DefaultStoryPointSupport implements StoryPointFieldSupport {
-
-    private static final Logger log = LoggerFactory.getLogger(DefaultStoryPointSupport.class);
 
     private final ScrumPokerSettings scrumPokerSettings;
     private final CustomFieldManager customFieldManager;
@@ -32,7 +27,6 @@ public class DefaultStoryPointSupport implements StoryPointFieldSupport {
 
     @Override
     public void save(String issueKey, Integer newValue) {
-        log.info("Saving estimation {} for issue {}...", newValue, issueKey);
         ApplicationUser applicationUser = context.getLoggedInUser();
         MutableIssue issue = issueService.getIssue(applicationUser, issueKey).getIssue();
         IssueInputParameters issueInputParameters = issueService.newIssueInputParameters();
@@ -40,13 +34,8 @@ public class DefaultStoryPointSupport implements StoryPointFieldSupport {
 
         IssueService.UpdateValidationResult validationResult = issueService.validateUpdate(applicationUser,
             issue.getId(), issueInputParameters);
-        if (validationResult.getErrorCollection().hasAnyErrors()) {
-            logErrors(issue.getKey(), validationResult.getErrorCollection());
-        } else {
-            IssueService.IssueResult updateResult = issueService.update(applicationUser, validationResult);
-            if (updateResult.getErrorCollection().hasAnyErrors()) {
-                logErrors(issue.getKey(), updateResult.getErrorCollection());
-            }
+        if (!validationResult.getErrorCollection().hasAnyErrors()) {
+            issueService.update(applicationUser, validationResult);
         }
     }
 
@@ -54,22 +43,13 @@ public class DefaultStoryPointSupport implements StoryPointFieldSupport {
         return NumberFormat.getInstance().format(newValue);
     }
 
-    private void logErrors(String issueKey, ErrorCollection errorCollection) {
-        log.error("Failed to update issue {} with errors:", issueKey);
-        if (errorCollection.hasAnyErrors()) {
-            errorCollection.getErrors().forEach((key, value) -> log.error("{}: {}", key, value));
-        }
-    }
-
     @Override
     public Integer getValue(String issueKey) {
         ApplicationUser applicationUser = context.getLoggedInUser();
         IssueService.IssueResult issueResult = issueService.getIssue(applicationUser, issueKey);
         if (issueResult.getErrorCollection().hasAnyErrors()) {
-            log.warn("Could not find value for issue {}.", issueKey);
             return null;
         }
-
         MutableIssue issue = issueResult.getIssue();
         return (Integer) issue.getCustomFieldValue(findStoryPointField());
     }